rename registries
