+++ src/aind_data_schema_models/organizations.py
@@ -5,7 +5,7 @@ from typing import Union
 from pydantic import BaseModel, ConfigDict, Field
 from typing_extensions import Annotated
 
-from aind_data_schema_models.registry import Registry, map_registry
+from aind_data_schema_models.registries import Registry, map_registry
 from aind_data_schema_models.utils import create_literal_class, read_csv
